nit: rename file from WidgetHelper.* to WidgetHelpers.*
